Account creation and account recovery module created
The account creation module was created to manage the registration of user accounts, with the following fields (first name, last name, date of birth, country, city, address, cell phone, zip code, email and password).
At the same time the account recovery module, located in the login, where the user's email is requested to manage the sending of an email with the code for the account recovery process.