Portfolio development part 1
The section for the purchase and sale of shares has been developed, with parameters of stock symbol, market price, quantity or amount depending on the type (shares or dollars), cost or estimated value and the type of order (Market order initially assigned).

In addition to the section for the review of the orders with a summary of the data.

On the other hand, there is an implementation of the state of the shares represented with a candlestick style graph, with the possibility of consulting the shares by means of a filter.

Translated with DeepL.com (free version)