Modification of user registration, connection for login and connection to buy and sell shares.
A modification was made to the user registry, adding two new sections (Employment Status) and (Financial Profile) to obtain other significant data for the creation of accounts.

On the other hand, the connection was created for login located in the “auth-service.service.ts”, to access the user's portfolio.

In addition to the connection to buy and sell shares, found in “trading-service.service.ts”, sending the fields (symbol, quantity, type of operation (buy / sell), type of order (in this case market) and validity of the order (in this case day).