independent sidebar and creation of user account funding.
The sidebar was adapted to be able to operate with all the other user options (portfolio, fund account, profile and logout).

On the other hand, the account funding section was created, in order to make the ach relationship and the subsequent transfer.

Likewise, a service was created for ach, to be able to operate with the endpoints of the backend.